@@ -15,77 +15,76 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.1.6",
-    "@angular/cdk": "^16.1.5",
-    "@angular/common": "^16.1.6",
-    "@angular/compiler": "^16.1.6",
-    "@angular/core": "^16.1.6",
-    "@angular/forms": "^16.1.6",
-    "@angular/platform-browser": "^16.1.6",
-    "@angular/platform-browser-dynamic": "^16.1.6",
-    "@angular/router": "^16.1.6",
-    "@angular/service-worker": "^16.1.6",
-    "@fullcalendar/core": "^6.1.8",
-    "@ngrx/effects": "^16.1.0",
-    "@ngrx/router-store": "^16.1.0",
-    "@ngrx/store": "^16.1.0",
-    "@ngrx/store-devtools": "^16.1.0",
+    "@angular/animations": "^17.1.0",
+    "@angular/cdk": "^17.2.1",
+    "@angular/common": "^17.1.0",
+    "@angular/compiler": "^17.1.0",
+    "@angular/core": "^17.1.0",
+    "@angular/forms": "^17.1.0",
+    "@angular/platform-browser": "^17.1.0",
+    "@angular/platform-browser-dynamic": "^17.1.0",
+    "@angular/router": "^17.1.0",
+    "@angular/service-worker": "^17.2.3",
+    "@fullcalendar/core": "^6.1.11",
+    "@ngrx/effects": "^17.1.1",
+    "@ngrx/router-store": "^17.1.1",
+    "@ngrx/store": "^17.1.1",
+    "@ngrx/store-devtools": "^17.1.1",
     "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^8.0.0",
-    "@types/lodash-es": "^4.17.8",
-    "bootstrap": "^5.3.0",
-    "crypto-js": "^4.1.1",
+    "@types/lodash-es": "^4.17.12",
+    "bootstrap": "^5.3.3",
+    "crypto-js": "^4.2.0",
     "hammerjs": "^2.0.8",
-    "husky": "^8.0.3",
-    "jquery": "^3.7.0",
+    "husky": "^9.0.11",
+    "jquery": "^3.7.1",
     "jszip": "^3.10.1",
     "ladda": "^2.0.3",
     "lodash-es": "^4.17.21",
     "masonry-layout": "^4.2.2",
-    "moment": "^2.29.4",
+    "moment": "^2.30.1",
     "ngx-logger": "^5.0.12",
-    "ngx-mask": "^16.2.5",
+    "ngx-mask": "^17.0.4",
     "ngx-masonry": "^14.0.1",
-    "ngx-toastr": "^17.0.2",
+    "ngx-toastr": "^18.0.0",
     "primeicons": "^6.0.1",
-    "primeng": "^16.0.2",
+    "primeng": "^17.11.0",
     "qs": "^6.11.2",
-    "rxjs": "~7.8.1",
+    "rxjs": "~7.8.0",
     "ts-md5": "^1.3.1",
-    "tslib": "^2.6.1",
-    "uuid": "^9.0.0",
-    "zone.js": "~0.13.1"
+    "tslib": "^2.3.0",
+    "uuid": "^9.0.1",
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.1.5",
-    "@angular-devkit/core": "^16.1.5",
-    "@angular-devkit/schematics": "^16.1.5",
-    "@angular-eslint/builder": "^16.1.0",
-    "@angular-eslint/eslint-plugin": "^16.1.0",
-    "@angular-eslint/eslint-plugin-template": "^16.1.0",
-    "@angular-eslint/schematics": "^16.1.0",
-    "@angular-eslint/template-parser": "^16.1.0",
-    "@angular/cli": "^16.1.5",
-    "@angular/compiler-cli": "^16.1.6",
-    "@angular/language-service": "^16.1.6",
-    "@types/bootstrap": "^5.2.6",
-    "@types/crypto-js": "^4.1.2",
-    "@types/jasmine": "~4.3.5",
-    "@types/uuid": "^9.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.2.0",
-    "@typescript-eslint/parser": "^6.2.0",
-    "eslint": "^8.45.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jsdoc": "^46.4.4",
+    "@angular-devkit/build-angular": "^17.2.3",
+    "@angular-devkit/core": "^17.2.2",
+    "@angular-devkit/schematics": "^17.2.2",
+    "@angular-eslint/builder": "^17.2.1",
+    "@angular-eslint/eslint-plugin": "^17.2.1",
+    "@angular-eslint/eslint-plugin-template": "^17.2.1",
+    "@angular-eslint/schematics": "^17.2.1",
+    "@angular-eslint/template-parser": "^17.2.1",
+    "@angular/cli": "^17.1.4",
+    "@angular/compiler-cli": "^17.1.0",
+    "@angular/language-service": "^17.2.3",
+    "@types/bootstrap": "^5.2.10",
+    "@types/crypto-js": "^4.2.2",
+    "@types/jasmine": "~5.1.0",
+    "@types/uuid": "^9.0.8",
+    "@typescript-eslint/eslint-plugin": "^7.1.0",
+    "@typescript-eslint/parser": "^7.1.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsdoc": "^48.2.0",
     "eslint-plugin-prefer-arrow": "^1.2.3",
     "jasmine-core": "~5.1.0",
-    "karma": "~6.4.2",
+    "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.1",
-    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "ng-packagr": "^16.1.0",
-    "typescript": "~5.1.6"
+    "ng-packagr": "^17.2.0",
+    "typescript": "~5.3.2"
   }
 }
