@@ -1,17 +1,17 @@
-import { ItemsEffects } from './store/effects/items.effects';
 import { LOCALE_ID, NgModule } from '@angular/core';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { AppRoutingModule } from './app-routing.module';
+import { AppRoutingModule } from './app.routing';
 import { AppComponent } from './app.component';
 import { CoreModule } from './core.module';
 import { SharedModule } from './shared/shared.module';
 import { StoreModule } from '@ngrx/store';
 import { environment } from 'src/environments/environment';
 import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-import { ROOT_REDUCERS } from './store/state/app.state';
+import { ROOT_REDUCERS } from './store/app.state';
 import { EffectsModule } from '@ngrx/effects';
 import { registerLocaleData } from '@angular/common';
 import localeEs from '@angular/common/locales/es';
+import { ROOT_EFFECTS } from '@store/app.root.effects';
 registerLocaleData(localeEs, 'es');
 @NgModule({
   declarations: [
@@ -24,7 +24,7 @@ registerLocaleData(localeEs, 'es');
     //StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: !isDevMode() }),
     // Effectos
     !environment.production ? StoreDevtoolsModule.instrument() : [],
-    EffectsModule.forRoot([ItemsEffects]),
+    EffectsModule.forRoot(ROOT_EFFECTS),
     CoreModule,
     SharedModule,
   ],
