@@ -1,7 +1,7 @@
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { EventEmitter, Injectable } from '@angular/core';
-
-import { BehaviorSubject, Observable, of } from 'rxjs';
+import { UpdateParams } from '@app/shared/common/utils';
+import { BehaviorSubject, Observable } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
@@ -20,8 +20,15 @@ export class AppTableService {
     private http: HttpClient,
   ) { }
 
-  getDatasourse(url: string): Observable<any> {
-    return this.http.get<any>(url);
+  getDatasourse(path: string): Observable<any> {
+    const objParams = {
+      start: 0,
+      length: 10,
+      search: '',
+      order: 'asc',
+    };
+    const { url, params } = UpdateParams(path, objParams);
+    return this.http.get<any>(url, { params });
   }
 
   habilitar(url: string, id: any, d: any): Observable<any> {
@@ -50,12 +57,13 @@ export class AppTableService {
     // }
     return this.http.get<any>(`${u}`);
   }
+  get(path: string): Observable<any> {
 
-  suggest(data: any, u: any, f: boolean): Observable<any> {
+    return this.http.get<any>(path);
+  }
 
-    if (f) {
-      return of(this.dataProds());
-    }
+  suggest(data: any, u: any): Observable<any> {
+
     const start = 0;
     const length = 10;
     const search = data ?? '';
@@ -66,42 +74,5 @@ export class AppTableService {
       .set('search', search)
       .set('order', order);
     return this.http.get<any>(`${u}`, { params: params });
-  }
-
-  dataProds(): any[] {
-    return [
-      {
-        id: 1,
-        title: `Fjallraven - Foldsack No. 1 Backpack, Fits 15 Laptops`,
-        price: 109.95,
-        codigo: 'l-001',
-        idUnidadMedida: 4,
-        description: `Your perfect pack for everyday use and walks in the forest.
-        Stash your laptop (up to 15 inches) in the padded sleeve, your everyday`,
-        category: `men's clothing`,
-        image: `https://fakestoreapi.com/img/81fPKd-2AYL._AC_SL1500_.jpg`,
-        rating: {
-          rate: 3.9,
-          count: 120,
-        },
-      },
-      {
-        id: 2,
-        title: `Mens Casual Premium Slim Fit T-Shirts `,
-        price: 22.3,
-        codigo: 't-001',
-        idUnidadMedida: 6,
-        description: `Slim-fitting style, contrast raglan long sleeve, three-button henley placket,
-         light weight & soft fabric for breathable and comfortable wearing.
-         And Solid stitched shirts with round neck made for durability and a
-         great fit for casual fashion wear and diehard baseball fans. The Henley style round neckline includes a three-button placket.`,
-        category: `men's clothing`,
-        image: `https://fakestoreapi.com/img/71-3HjGNDUL._AC_SY879._SX._UX._SY._UY_.jpg`,
-        rating: {
-          rate: 4.1,
-          count: 259,
-        },
-      },
-    ];
   }
 }
