@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { PrimengModule } from './primeng/primeng.module';
+import { AppPrimengModule } from './primeng/primeng.module';
 import { AppHeaderModule } from './components/app-header/app-header.module';
 import { AppSidebarModule } from './components/app-sidebar/app-sidebar.module';
 import { AppInputModule } from './forms/app-input/app-input.module';
@@ -14,6 +14,8 @@ import { UnspscModule } from './components/unspsc/unspsc.module';
 import { AppFinderModule } from './forms/app-finder/app-finder.module';
 import { AppHomeModule } from './components/app-home/app-home.module';
 import { AppTotalModule } from './components/app-total/app-total.module';
+import { AppTableGridModule } from './components/app-table-grid/app-table-grid.module';
+import { AppSelectRolesModule } from './components/app-select-roles/app-select-roles.module';
 
 @NgModule({
   imports: [
@@ -23,7 +25,7 @@ import { AppTotalModule } from './components/app-total/app-total.module';
     AppClickOutsideDirective,
   ],
   exports: [
-    PrimengModule,
+    AppPrimengModule,
     AppInputModule,
     AppInputFileModule,
     AppHeaderModule,
@@ -31,11 +33,13 @@ import { AppTotalModule } from './components/app-total/app-total.module';
     AppSelectModule,
     AppFinderModule,
     AppTableModule,
+    AppTableGridModule,
     AppTotalModule,
     AppToolModule,
     AppUbigeoModule,
     UnspscModule,
     AppHomeModule,
+    AppSelectRolesModule,
     //
     AppClickOutsideDirective,
   ],
