@@ -7,8 +7,7 @@ import { Component, OnInit } from '@angular/core';
 })
 export class HomeAdministracionComponent implements OnInit {
 
-  constructor() {
-  }
+  constructor() { }
 
   ngOnInit(): void {
     if (1) { }
