@@ -2,54 +2,26 @@ import { Component, Injector, OnInit } from '@angular/core';
 import { UntypedFormGroup, Validators } from '@angular/forms';
 import { AbstractDocument } from '@app/shared/common/classes';
 import { AppTable } from '@app/shared/components/app-table/app-table.interface';
-
+import { ConfirmationService } from 'primeng/api';
 @Component({
   selector: 'app-usuario',
   templateUrl: './usuario.component.html',
   styleUrls: ['./usuario.component.scss'],
+  providers: [ConfirmationService],
 })
 export class UsuarioComponent extends AbstractDocument implements OnInit {
 
   fullPath = '/usuario';
   headers: AppTable[] = [
-    {
-      field: 'id',
-      label: 'id',
-      visible: false,
-    },
-    {
-      field: 'nombres',
-      label: 'Nombres',
-    },
-    {
-      field: 'apellidos',
-      label: 'Apellidos',
-    },
-    {
-      field: 'email',
-      label: 'Correo',
-    },
-    {
-      field: 'username',
-      label: 'usuario',
-    },
-    {
-      field: 'password',
-      label: 'password',
-      visible: false,
-    },
-    {
-      field: 'telefono',
-      label: 'telefono',
-    },
-    {
-      field: 'habilitado',
-      label: 'estado',
-      estado: true,
-    },
-
+    { field: 'id', label: 'id', visible: false },
+    { field: 'nombres', label: 'Nombres' },
+    { field: 'apellidos', label: 'Apellidos' },
+    { field: 'email', label: 'Correo' },
+    { field: 'username', label: 'usuario' },
+    { field: 'password', label: 'password', visible: false },
+    { field: 'telefono', label: 'telefono' },
+    { field: 'habilitado', label: 'habilitado', type: 'habilitado' },
   ];
-  data: any[] = [];
 
   form: UntypedFormGroup = this.fb.group({
     id: [],
@@ -63,9 +35,11 @@ export class UsuarioComponent extends AbstractDocument implements OnInit {
     habilitado: [true],
     telefono: [],
     photo: [],
+    id_rol: ['', [Validators.required]],
   });
   constructor(
     injector: Injector,
+    private confirmationService: ConfirmationService,
   ) {
     super(injector);
   }
@@ -75,4 +49,21 @@ export class UsuarioComponent extends AbstractDocument implements OnInit {
     this.hashValues = ['password', 'rep_password'];
   }
 
+  selectChange(e: any): void {
+    if (e?.id === 1) {
+      this.confirmationService.confirm({
+        message: `Seguro que deseas agregar el rol de Super Admin?
+        <br>
+        <a href="/configuracion/registro-de-rol" class="text-info float-end mt-3" rel="noopener noreferrer">Agregar Rol</a>`,
+        header: 'Warning',
+        icon: 'pi pi-exclamation-triangle text-warning',
+        accept: () => {
+          this.form.patchValue({ id_rol: e.id });
+        },
+        reject: () => {
+          this.form.patchValue({ id_rol: null });
+        },
+      });
+    }
+  }
 }
