@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Store } from '@ngrx/store';
+import { AppStateStore } from '@store/app.state';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'app-caja',
@@ -7,7 +10,9 @@ import { Component, OnInit } from '@angular/core';
 })
 export class CajaComponent implements OnInit {
 
-  constructor() { }
+  loading$: Observable<boolean> = new Observable();
+
+  constructor(private store: Store<AppStateStore>) { }
 
   ngOnInit(): void {
     if (1) { }
