@@ -4,7 +4,7 @@ import { CommonModule } from '@angular/common';
 import { InicioModule } from './inicio/inicio.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import { ViewsRoutingModule } from './views-routing.module';
+import { ViewsRoutingModule } from './views.routing';
 import { ViewsComponent } from './views.component';
 import { PlanillasModule } from './planillas/planillas.module';
 import { CajaModule } from './caja/caja.module';
